Move `slice::Chunks*` to a new submodule
